# Conflicts:
#	ProgAvanceeProjectWPF/DAO/CategoryDAO.cs
#	ProgAvanceeProjectWPF/DAO/MemberDAO.cs
#	ProgAvanceeProjectWPF/Metier/Member.cs
#	ProgAvanceeProjectWPF/Metier/Ride.cs
#	ProgAvanceeProjectWPF/Pages/Responsibles/ManagmentRide.xaml.cs
#	ProgAvanceeProjectWPF/Pages/Responsibles/ResponsibleHub.xaml.cs

+++ ProgAvanceeProjectWPF/Metier/Member.cs
@@ -5,12 +5,16 @@ using System.Linq;
 public class Member : Person
 {
     private double balance = 0;
+    private DateTime datePayment;
+    private bool paymentCheck;
     private List<Category> categories = new List<Category>();
     private List<Bike> bikes = new List<Bike>();
     private List<Inscription> inscriptions = new List<Inscription>();
     private List<Vehicle> vehicles = new List<Vehicle>();
+    private List<Message> messages = new List<Message>();
 
     AbstractDAOFactory adf = AbstractDAOFactory.GetFactory(DAOFactoryType.MS_SQL_FACTORY);
+    MemberDAO dao = new MemberDAO();
 
     public Member()
     {
@@ -22,10 +26,29 @@ public class Member : Person
         this.balance = balance;
     }
 
+    public Member(Guid id, string name, string firstName, string tel, string login, string passWord, double balance, DateTime datePayment, bool paymentCheck) : base(id, name, firstName, tel, login, passWord)
+    {
+        this.balance = balance;
+        this.datePayment = datePayment;
+        this.paymentCheck = paymentCheck;
+    }
+
     public double Balance
     {
         get { return balance; }
         set { balance = value; }
+    }
+
+    public DateTime DatePayment
+    {
+        get { return datePayment; }
+        set { datePayment = value; }
+    }
+
+    public bool PaymentCheck
+    {
+        get { return paymentCheck; }
+        set { paymentCheck = value; }
     }
 
     public List<Category> Categories
@@ -45,6 +68,12 @@ public class Member : Person
         get { return inscriptions; }
         set { inscriptions = value; }
     }
+    
+    public List<Message> Messages
+    {
+        get { return messages; }
+        set { messages = value; }
+    }
 
     public List<Vehicle> Vehicles
     {
@@ -62,48 +91,58 @@ public class Member : Person
         return balance <= amount;
     }
 
+
+    //Récupère tous les membres
     public List<Member> GetAllMembers()
     {
         MemberDAO dao = new MemberDAO();
         return dao.GetAllMembers();
     }
 
-    public bool AddMember(string AddName, string AddFirstName, string AddTelephone, string AddLogin, string AddPassWord, double AddBalance)
+    //CRUD MEMBER
+    public bool AddMember(string AddName, string AddFirstName, string AddTelephone, string AddLogin, string AddPassWord, double AddBalance, Category cat)
     {
         DAO<Member> memberDAO = adf.GetMemberDAO();
 
         Member m = new Member(Guid.NewGuid(), AddName, AddFirstName, AddTelephone, AddLogin, AddPassWord, AddBalance);
 
-        return memberDAO.Create(m);
+        if (memberDAO.Create(m))
+        {
+            m.AddCategory(cat, m);
+            return true;
+        }
+        else
+        {
+            return false;
+        }
     }
-
-    public bool UpdateMember(Member m, string UpdateName, string UpdateFirstName, string UpdateTelephone, string UpdateLogin, string UpdatePassWord, double UpdateBalance)
+    public bool UpdateMember(Member m)
     {
         DAO<Member> memberDAO = adf.GetMemberDAO();
-
-        m.Name = UpdateName;
-        m.FirstName = UpdateFirstName;
-        m.Tel = UpdateTelephone;
-        m.Login = UpdateLogin;
-        m.PassWord = UpdatePassWord;
-        m.Balance = UpdateBalance;
-
         return memberDAO.Update(m);
     }
-
     public bool DeleteMember(Member m)
     {
         DAO<Member> memberDAO = adf.GetMemberDAO();
 
-        return memberDAO.Delete(m);
+        if (dao.RemoveMemberCategory(m))
+        {
+            memberDAO.Delete(m);
+            return true;
+        }
+        else
+        {
+            return false;
+        }
     }
 
+    //Vérifie si le Membre existe en Base de données
     public Member loginCheck(string login, string password)
     {
-        MemberDAO dao = new MemberDAO();
         return dao.loginCheck(login, password);
     }
 
+    //Ajout modification Suppression d'un vélo 
     public void AddBike(Bike bike)
     {
         this.bikes.Add(bike);
@@ -129,10 +168,28 @@ public class Member : Person
     }
     public bool AddCategory(Category category, Member member)
     {
-        MemberDAO dao = new MemberDAO();
-        dao.AddMemberCategory(category, member);
-        this.Categories.Add(category);
-        return true;
+        if(dao.AddMemberCategory(category, member))
+        {
+            this.Categories.Add(category);
+            return true;
+        }
+        else
+        {
+            return false;
+        }
+    }
+
+    public List<Member> CheckDate(List<Member> members)
+    {
+        foreach (Member m in members)
+        {
+            if (m.DatePayment.AddYears(1) < DateTime.Now)
+            {
+                m.PaymentCheck = false;
+                m.UpdateMember(m);
+            }
+        }
+        return members;
     }
